// src/auth/enum/errors.enum.ts
export enum Errors {
  VALIDATION_REQUIRED = 'VALIDATION_REQUIRED',
  VALIDATION_LOGIN = 'AUTH_VALIDATION_LOGIN',
  VALIDATION_NAME = 'AUTH_VALIDATION_NAME',
  VALIDATION_PASSWORD = 'AUTH_VALIDATION_PASSWORD',
  VALIDATION_EMAIL = 'AUTH_VALIDATION_EMAIL',
  VALIDATION_CONTACTS = 'AUTH_VALIDATION_CONTACTS',
  USERNAME_ALREADY_EXISTS = 'SIGN_UP_USERNAME_ALREADY_EXISTS',
  COULDNT_FOUND_USER = 'LOGIN_COULDNT_FOUND_USER',
  UNCORRECT_LOGIN_DATA = 'LOGIN_UNCORRECT_LOGIN_DATA',
  UNAUTHORIZED = 'AUTH_UNAUTHORIZED',
}